+++ pywbemtools/pywbemcli/_click_extensions.py
@@ -3,7 +3,10 @@ This file contains extensions to Click specifically for pywbemcli
 """
 import sys
 from collections import OrderedDict
+
 import click
+
+from ._common import GENERAL_OPTS_TXT
 
 
 class PywbemcliGroup(click.Group):
@@ -14,6 +17,12 @@ class PywbemcliGroup(click.Group):
         appears in the source code for each command group.
     This extension has a general name because it may be used for more than
     one extension to the Click.Group class.
+
+    2. Force use of our method pywbemcli_format_usage rather than
+       the click define Command.format_usage()
+
+    It is used on all group decorators (@click.group(...)) except for the
+    top level in pywbemcli which uses the PywbemcliTopGroup class.
     """
 
     # Use ordered dictionary to sort commands by their order defined in the
@@ -23,16 +32,19 @@ class PywbemcliGroup(click.Group):
         Use OrderedDict to keep order commands inserted into command dict.
         Only required for Python versions that do not order dictionaries.
         Must be set after calling superclass __inits_ because click forces
-        {} even if user were to set commands to OrderedDict
+        {} for this variable even if user were to set commands to OrderedDict
         """
         super(PywbemcliGroup, self).__init__(name, commands, **attrs)
-        #: the registered subcommands by their exported names.
+
         if sys.version_info < (3, 6):
             self.commands = commands or OrderedDict()
 
+        # Replace Click.Command.format_usage with local version
+        click.core.Command.format_usage = pywbemcli_format_usage
+
     def list_commands(self, ctx):
         """
-        Replace list_commands to eliminate the sort.
+        Replace click.list_commands to eliminate the sort of cmd names.
         """
         return self.commands.keys()
 
@@ -55,11 +67,15 @@ class PywbemcliTopGroup(click.Group):
     def list_commands(self, ctx):
         """
         Order The top level commands by sorting and then moving any commands
-        defined in  move_to_end list to the end of the list.
+        defined in  move_to_end list to the end of the list.  This is
+        the class override for click.Group ONLY with the top group, the
+        click startup group in pywbemcli because in reorders the commands
+        display specifically as we want them for pywbemcli.
         """
         # tuple of commands to move to bottom after sort
         move_to_end = ('connection', 'help', 'repl')
 
+        # Sort because thier is no particular order for the groups
         cmd_list = sorted(self.commands.keys())
         pop_count = 0
         # reorder list so the move_to_end list commands are at bottom
@@ -69,23 +85,38 @@ class PywbemcliTopGroup(click.Group):
                 pop_count += 1
         return cmd_list
 
+
 class PywbemcliCommand(click.Command):
     """
     Modify the command usage formatter to show the commands in a format
-    that fits how we use the tool
+    that fits how we use the tool. This should be the cls= argument
+    for all pywbemcli command decorators (i.e. `@class_group.command(`).
     """
 
     def format_usage(self, ctx, formatter):
         """
+        Replaces click.Command.format_usage
         """
         pieces = self.collect_usage_pieces(ctx)
         cmd_paths = ctx.command_path.split()
+        # Reorder pieces to order GENERAL_OPTS_TXT, <cmd> ...
+        new_pieces = [GENERAL_OPTS_TXT] + cmd_paths[1:] + pieces[1:] \
+            + [pieces[0]]
+        formatter.write_usage(cmd_paths[0], " ".join(new_pieces))
 
-        # If cmd_path has multiple components we are showing
-        # app name, CMDGRP <cmd>. Break out CMDGRP and <cmd>
-        # and reinsert after pieces[0] (the app name)
-        if len(cmd_path) > 1:
-            new_pieces = [pieces[0]] + cmd_paths[1:] + pieces[1:]
-            formatter.write_usage(cmds[0], " ".join(new_pieces))
-        else:
-            formatter.write_usage(ctx.command_path, " ".join(pieces))
+
+def pywbemcli_format_usage(self, ctx, formatter):
+    """
+    Replaces click.Command.format_usage in the fixes for click.Group
+    """
+    pieces = self.collect_usage_pieces(ctx)
+    cmd_paths = ctx.command_path.split()
+
+    # If cmd_path has multiple components we are showing
+    # app name, CMDGRP <cmd>. Break out CMDGRP and <cmd>
+    # and reinsert after pieces[0] (the app name)
+    if len(cmd_paths) > 1:
+        new_pieces = [pieces[0]] + cmd_paths[1:] + pieces[1:]
+        formatter.write_usage(cmd_paths[0], " ".join(new_pieces))
+    else:
+        formatter.write_usage(ctx.command_path, " ".join(pieces))